filesystem: attempt to read each superblock
Updated the `filesystem.Probe` function to attempt to read each
superblock and return only when a valid superblock is found. This
accounts for the case where there's an error reading an earlier
superblock but success reading a later superblock.

See https://github.com/siderolabs/omni/discussions/1633 for more
details.

Thank you @lgfa29 and @JustinAzoff for the assistance in debugging this!